_passive_income/ArmyOfRobots/My1strobot/requirements.txt
PyQt5
pandas
PyYAML
alpaca-trade-api
ta
yfinance
